remove plugins to composables
